MINOR: Improve GroupCoordinatorService#onPartitionsDeleted path (#21055)
The `GroupCoordinatorService#onPartitionsDeleted` path schedules
operations to delete committed offsets, waits on all the results,
schedules operations to delete share state, and waits on the results.
This is inefficient and holds the metadata publisher for longer than
needed. This patch refactors the code to schedule all the operations for
both and then waits on all the results. It relies on existing tests.

Reviewers: Sean Quah <squah@confluent.io>, Lianet Magrans <lmagrans@confluent.io>