KAFKA-19944: kafka-share-groups.sh describe offsets message improvement (#21024)
kafka-share-groups.sh --describe --offsets prints an empty table
consisting only of headers when there is no offset information to
display for a share group. This is most likely only when a topic which
was formerly being consumed is deleted. This PR displays a message
instead of an empty table, similar to how the tool works if the list of
members is requested and there are no active members.

Reviewers: Apoorv Mittal <apoorvmittal10@gmail.com>