KAFKA-19439 OffsetFetch API does not return group level errors correctly with version 1 (#21063)
The OffsetFetch API does not support top level errors in version 1.
Hence, the top level error must be returned at the partition level.

Side note: It is a tad annoying that we create error response in
multiple places (e.g. KafkaApis, Group CoordinatorService). There were a
reason for this but I cannot remember.

Reviewers: PoAn Yang <payang@apache.org>, Dongnuo Lyu
<dlyu@confluent.io>, Sean Quah <squah@confluent.io>, Ken Huang
<s7133700@gmail.com>, TengYao Chi <kitingiao@gmail.com>, Chia-Ping Tsai
<chia7712@gmail.com>

---------

Co-authored-by: David Jacot <david.jacot@gmail.com>