KAFKA-19956: Fix flaky testRemoteAssignorRange (#21060)
This has been flaky, failing with "Topic does not exist" received in a
response to a commit.

This could happen given that the test had the default auto-commit
enabled and was not explicitly closing the consumer (consumer.close
triggered as part of the test infra tear down, which does close with
timeout 0). Seems the topic could end up deleted before the auto-commit
on close was processed.

Disabling auto-commit, not needed for the test.

Reviewers: Andrew Schofield <aschofield@confluent.io>