KAFKA-19957: Accumulate offsets response in same batch (#21061) (#21064)
Merges contiguous AcquiredRecords with the same delivery count into
single records.

The PR accumulates multiple offsets in a single batch only if the
offsets are part of same in-flight batch i.e. if cache has 0-4, and 5-9
and the records acquired are 3,4,5,6, with same delivery count, then
output shall be [3-4],[5-6].  This has been done to make sure that
offsets accumulation is not past  the `batch size` in request. The
in-flight batches originally formed on  the broker are created as per
the `batch size` from the client hence  accumulating from single batch
is safe.

However, if in a same share group multiple clients request with
different batch sizes then there would be discrepancy. Having said that,
the similar problem might occur with complete batch acquisition in this
case hence  that should be fixed separately, if needed.

Reviewers: Andrew Schofield <aschofield@confluent.io>