KAFKA-19945: Always set status field in StreamsGroupHeartbeat response (#21031)
The status field should always set in the response, even when empty, to
ensure that the status gets reset on the client.

We are not doing this due to an oversight - it is defined as nullable,
and it is null when not set. So status does not clear correctly on the
client, which ignores the field if it's null.

This can cause the streams application to incorrectly timeout, if the
source topic does not exist when the application is first started, but
it is created after the application started. Otherwise, there is no
noticable difference.

Reviewers: Lucas Brutschy <lbrutschy@confluent.io>