KAFKA-19678: Refactor to reset timestamp to zero when no open iterators (#21091)
This PR changes the approach for handling the case when there are no
open iterators for the `oldest-iterator-open-since-ms` metric.
Previously, we would register the metric when the store opens its first
iterator, then deregister the metric when it closes the last remaining
open iterator.  Under heavy load this process can cause performance
degradation as detailed in the Jira ticket
https://issues.apache.org/jira/browse/KAFKA-19678.  This PR updates the
handling of the metric to register and deregister on store opening and
closing respectively, and handle the case of current open iterators by
returning a `0` to indicate this state.

Reviewers: Matthias Sax <mjsax@apache.org>, Ramesh Sinha