KAFKA-19978: Don't enter running state as long as group is not ready (#21110)
Since internal topics are created asynchronous, and we do not want to
have heartbeats failing, when an application has not yet created its
internal topics, it remains in NOT_READY, until the internal topics are
created.

The way the client-side is implemented right now, the Kafka Streams
application will still  enter RUNNING state with an empty assignment,
when internal topics are missing. From the outside, it will therefore
look like the initial set-up of the application has finished, while the
application is still waiting for the group to become ready.

This propagates the state of the group as part of the assignment, so
that  we delay entering a "RUNNING" state until the group has become
ready.

Reviewers: Matthias J. Sax <matthias@confluent.io>