KAFKA-19749 The `version-mapping` of kafka-features.sh should work without requiring the bootstrap (#20619)
# Description
In kafka-features.sh, exclusive argument group of `--bootstrap-server`
and `--bootstrap-controller` is required in the main parser, but
`version-mapping` and `feature-dependencies` subcommands don't need the
information.

# Changes
To avoid change in argument level, the exclusive group is not moved. It
is now not required and only check for existence if the subcommand needs
it.

Since `FeatureCommand#mainNoExit` catches `ArgumentParserException`,
`ArgumentParser#parseArgs` is used instead of `parseArgsOrFail`. It
makes testing easier.

Two tests are added. One tests whether not-remote subcommand can execute
without bootstrap. Another tests whether remote subcommand cannot
execute without bootstrap.

Reviewers: Ken Huang <s7133700@gmail.com>, Jhen-Yung Hsu
 <jhenyunghsu@gmail.com>, TaiJuWu <tjwu1217@gmail.com>, Chia-Ping Tsai
 <chia7712@gmail.com>