KAFKA-19026: AlterConfigPolicy incompatibility between ZK mode and KR… (#19263)
…aft mode when using AlterConfigOp.OpType.SUBTRACT

Modified ZkAdminManager.scala so that on OpType.SUBTRACT the policy
receives the modified configs, as happens in KRaft mode.

This similarly fixes the OpType.APPEND differences.

Note that the policy behavior on OpType.DELETE is different when
altering Broker and Topic resources. For topics the policy does not see
a map entry, for brokers the config value is null. This was the existing
behavior for KRaft and this commit does not change that.

ClusterTest added.