KAFKA-15599: Move DefaultExternalKRaftMetrics to the server module (#21163)
Another small step in moving all of `kafka.raft` classes out of the
`core` module.

The `DefaultExternalKRaftMetrics` class needs:
- `ControllerMetadataMetrics` from `metadata`
- `ExternalKRaftMetrics` from `raft`
-  `BrokerServerMetrics` from `server`.

The `server` module already depends on `metadata` and `raft`. It also
already has a bunch of metrics related classes in
`org.apache.kafka.server.metrics`, so it seems the best fit.

Reviewers: Vincent Jiang
 <84371940+vincent81jiang@users.noreply.github.com>, Chia-Ping Tsai
 <chia7712@gmail.com>